+++ src/elements/Modal/Modal.fixture.jsx
@@ -7,7 +7,7 @@
 import React from 'react'
 
 import Modal from './Modal'
-// import FormGroup from 'forms/FormGroup'
+import FormGroup from 'forms/FormGroup'
 import FormButton from 'forms/FormButton'
 // import Menu from 'elements/Menu'
 
@@ -23,17 +23,18 @@ class Simple extends React.PureComponent /** @lends @PopupsContainer.prototype *
   constructor(props) {
     super(props)
     this.state = {
-      show: true,
+      open: true,
     }
-    // setTimeout(this.hide, 3000)
+    // setTimeout(this.close, 3000)
     // setInterval(this.toggle, 5000)
   }
-  show = () => this.setState({ show: true })
-  hide = () => this.setState({ show: false })
+  open = () => this.setState({ open: true })
+  close = () => this.setState({ open: false })
+  handleOpenState = ({ open }) => this.setState({ open })
   toggle = () => {
     this.modalRef && this.modalRef.toggle()
-    // this.setState(({ show }) => {
-    //   return { show: !show }
+    // this.setState(({ open }) => {
+    //   return { open: !open }
     // })
   }
   setModalRef = (ref) => {
@@ -41,38 +42,56 @@ class Simple extends React.PureComponent /** @lends @PopupsContainer.prototype *
   }
   renderActions() {
     return (
-      <FormButton
-        id="ok"
-        icon="faCheck"
-        text="Ok"
-        theme="primary"
-        inline
-      />
+      <FormGroup id="Actions" flow fullWidth>
+        <FormGroup id="ActionsMain" flow>
+          <FormButton
+            id="ok"
+            icon="faCheck"
+            text="Ok"
+            theme="primary"
+            inline
+          />
+        </FormGroup>
+        <FormGroup id="ActionsRight" flow align="right">
+          <FormButton
+            id="cancel"
+            icon="faTimes"
+            text="Cancel"
+            inline
+          />
+        </FormGroup>
+      </FormGroup>
     )
   }
+  onAction = ({ id }) => {
+    console.log('onAction', id)
+  }
   render() {
-    const { show } = this.state
+    const { open } = this.state
     const actions = this.renderActions()
     // const isElement = React.isValidElement(actions)
     return (
       <React.Fragment>
         <Modal
           id="simple"
-          icon="faExclamationCircle"
-          show={show}
+          theme="info"
+          // icon="faExclamationCircle"
+          open={open}
           className="ModalClass"
           windowClassName="WindowClass"
           wrapperClassName="WrapperClass"
-          onClose={this.hide}
+          onAction={this.onAction}
+          onClose={this.close}
           onOpen={this.open}
+          handleOpenState={this.handleOpenState}
           ref={this.setModalRef}
           actions={actions}
           title="Modal title"
           content="Main content"
-          theme="error"
-          // width="md"
+          width="sm"
           // leftContent="left" // Left column (with icon visual, eg)
           showCloseButton
+          autoClose
         >
           <div className="simpleContainer">
             <div className="simpleStub">simpleStub</div>